+++ hdf/test/vers.c
@@ -38,7 +38,7 @@ void test_vers()
   RESULT("Hgetlibversion");
   MESSAGE(5,printf("Library Version\n"););
   MESSAGE(5,printf("---------------\n"););
-  MESSAGE(5,printf("Major:\t\t%d\nMinor:\t\t%d\nRelease:\t%d\nString:\t\t\"%s\"\n",lmajor, lminor, lrelease, lstring););
+  MESSAGE(5,printf("Major:\t\t%u\nMinor:\t\t%u\nRelease:\t%u\nString:\t\t\"%s\"\n",(unsigned)lmajor, (unsigned)lminor, (unsigned)lrelease, lstring););
 
 }
